add CompleteLinksToReferencesWorker test
